+++ packages/clerk-js/sandbox/app.ts
@@ -32,6 +32,7 @@ const AVAILABLE_COMPONENTS = [
   'waitlist',
   'pricingTable',
   'apiKeys',
+  'configureSSO',
   'oauthConsent',
   'taskChooseOrganization',
   'taskResetPassword',
@@ -136,6 +137,7 @@ const componentControls: Record<AvailableComponent, ComponentPropsControl> = {
   waitlist: buildComponentControls('waitlist'),
   pricingTable: buildComponentControls('pricingTable'),
   apiKeys: buildComponentControls('apiKeys'),
+  configureSSO: buildComponentControls('configureSSO'),
   oauthConsent: buildComponentControls('oauthConsent'),
   taskChooseOrganization: buildComponentControls('taskChooseOrganization'),
   taskResetPassword: buildComponentControls('taskResetPassword'),
@@ -468,6 +470,9 @@ void (async () => {
     '/api-keys': () => {
       Clerk.mountAPIKeys(app, componentControls.apiKeys.getProps() ?? {});
     },
+    '/configure-sso': () => {
+      Clerk.__experimental_mountConfigureSSO(app, componentControls.configureSSO.getProps() ?? {});
+    },
     '/oauth-consent': () => {
       const searchParams = new URLSearchParams(window.location.search);
       const scopes = (searchParams.get('scope')?.split(',') ?? []).map(scope => ({
@@ -481,7 +486,6 @@ void (async () => {
           scopes,
           oauthClientId: 'Wg9fP2d0pSFXCZ1u',
           redirectUrl: searchParams.get('redirect_uri') ?? 'http://localhost:4000/oauth/callback',
-          __internal_enableOrgSelection: true,
         },
       );
     },