+++ packages/clerk-js/src/core/query-core.ts
@@ -1,3 +1,0 @@
-import { QueryClient } from '@tanstack/query-core';
-
-export { QueryClient };