+++ packages/ui/src/internal/appearance.ts
@@ -672,6 +672,16 @@ export type ElementsConfig = {
   enterpriseConnectionButton: WithOptions;
   enterpriseConnectionButtonText: WithOptions;
 
+  configureSSOWizardHeader: WithOptions;
+  configureSSOWizardHeaderItem: WithOptions<string, ActiveState>;
+  configureSSOWizardHeaderItemBullet: WithOptions<string, ActiveState>;
+  configureSSOWizardHeaderItemLabel: WithOptions<string>;
+  configureSSOWizardHeaderSeparator: WithOptions;
+  configureSSOWizardBody: WithOptions<string>;
+  configureSSOWizardStepIndicator: WithOptions;
+  configureSSOWizardFooterPreviousButton: WithOptions;
+  configureSSOWizardFooterContinueButton: WithOptions;
+
   web3SolanaWalletButtonsRoot: WithOptions;
   web3SolanaWalletButtons: WithOptions;
   web3SolanaWalletButtonsIconButton: WithOptions<string, LoadingState>;
@@ -1014,6 +1024,7 @@ export type CheckoutTheme = Theme;
 export type PlanDetailTheme = Theme;
 export type SubscriptionDetailsTheme = Theme;
 export type APIKeysTheme = Theme;
+export type __experimental_ConfigureSSOTheme = Theme;
 export type OAuthConsentTheme = Theme;
 export type TaskChooseOrganizationTheme = Theme;
 export type TaskResetPasswordTheme = Theme;
@@ -1090,6 +1101,10 @@ export type Appearance<T = Theme> = T &
      * Theme overrides that only apply to the `<APIKeys />` component
      */
     apiKeys?: T;
+    /**
+     * Theme overrides that only apply to the `<ConfigureSSO />` component
+     */
+    __experimental_configureSSO?: T;
     /**
      * Theme overrides that only apply to the `<OAuthConsent />` component
      */